Added some usecases to handle

+++ src/cli/index.ts
@@ -18,45 +18,58 @@ const QUESTIONS = {
     }
 };
 
-function run() {
-    prompt([QUESTIONS.SCAN]).then(({ scanPath }) => {
-        const resolvedPath = resolve(process.cwd(), scanPath as string);
+function run(): void {
+	prompt([QUESTIONS.SCAN])
+		.then(({ scanPath }) => {
+			const resolvedPath = resolve(process.cwd(), scanPath as string);
 
-        if (!existsSync(resolvedPath)) {
-            throw new Error('Directory does not exist, please pass a valid path.');
-        }
+			if (!existsSync(resolvedPath)) {
+				throw new Error('Directory does not exist, please pass a valid path.');
+			}
 
-        const findings = getFiles(resolvedPath);
-        console.table(findings);
+			const findings = getFiles(resolvedPath);
 
-        prompt([QUESTIONS.WRITE]).then(({ writePath }) => {
-            const filePath: string = resolve(process.cwd(), writePath as string);
+			if (!Object.keys(findings).length) {
+				console.log('No duplicates where found.');
+				return;
+			}
 
-            if (filePath.endsWith('.js')) {
-                const data: Array<string> = [];
+			console.table(findings);
 
-                Object.keys(findings).forEach((finding, i) => {
-                    data.push(`const string${i} = '${finding}';`);
-                });
+			prompt([QUESTIONS.WRITE])
+				.then(({ writePath }) => {
+					const filePath: string = resolve(process.cwd(), writePath as string);
 
-                writeFileSync(filePath, data.join('\n'), 'utf8');
-                return;
-            }
+					if (filePath.endsWith('.js')) {
+						const data: Array<string> = [];
 
-            if (filePath.endsWith('.json')) {
-                const data: { [key: string]: string } = {};
+						Object.keys(findings).forEach((finding, i) => {
+							data.push(`const string${i} = '${finding}';`);
+						});
 
-                Object.keys(findings).forEach((finding, i) => {
-                    data[`string${i}`] = `${finding}`;
-                });
+						writeFileSync(filePath, data.join('\n'), 'utf8');
+						return;
+					}
 
-                writeFileSync(filePath, JSON.stringify(data, null, 2), 'utf8');
-                return;
-            }
+					if (filePath.endsWith('.json')) {
+						const data: { [key: string]: string } = {};
 
-            throw new Error(`File format not supported`);
-        });
-    });
+						Object.keys(findings).forEach((finding, i) => {
+							data[`string${i}`] = `${finding}`;
+						});
+
+						writeFileSync(filePath, JSON.stringify(data, null, 2), 'utf8');
+						return;
+					}
+
+					throw new Error(`File format not supported`);
+				})
+				.catch((e) => {
+					console.error(e);
+				});
+		}).catch((e) => {
+			console.error(e);
+		});
 }
 
 program
